==> Parking_Project/2_Google Maps API - Carpark Coordinates.ipynb <==
# Google Maps API

This is a simple program to use the Google Maps API to obtain the geo coordinates of each carpark address and append them to the original dataframe. The result is saved into a csv file.

import pandas as pd
from math import radians
import pandas as pd
from haversine import haversine

### Load csv and set up variables

# Read carpark csv as dataframe
df = pd.read_csv('Parking.csv')


# Set up variables
front_url = "https://maps.googleapis.com/maps/api/geocode/json?address="
end_url = "&components=country:SG&key=AIzaSyB-lR8VoOizlVvhK-p8CR6Lol-wb2RgSM0"
lat_list = []
lng_list = []

df['postal'] = df['Carpark Address'].str[-7:-1]

### Call Google Maps API, append results to dataframe and save to csv

i = 1

for index, row in df.iterrows(): 
    try:
        url = front_url + row['postal'] + end_url
        address = pd.read_json(url)
        lat = address['results'][0]['geometry']['location']['lat']
        lng = address['results'][0]['geometry']['location']['lng']
        lat_list.append(radians(lat)) # saves latitude in radians
        lng_list.append(radians(lng)) # saves longitude in radians
    
    except:
        try:
            url = front_url + row['Carpark Address'].replace(' ', '+') + end_url
            address = pd.read_json(url)
            lat = address['results'][0]['geometry']['location']['lat']
            lng = address['results'][0]['geometry']['location']['lng']
            lat_list.append(radians(lat)) # saves latitude in radians
            lng_list.append(radians(lng)) # saves longitude in radians
        except:
            print(row['Carpark Address'])
            lat_list.append('NAN')
            lng_list.append('NAN')
        
    print(str(i) + " of " + str(len(df)))
    i += 1
        
# Create columns in dataframe
df['Latitude_rad'] = lat_list
df['Longitude_rad'] = lng_list
df['Coord_rad'] = list(zip(df.Latitude_rad, df.Longitude_rad))


# Save to csv
df.to_csv('Parking_withcoords_2.csv', index=False)
